+++ tests/leb-freeze/test_fp.py
@@ -2,7 +2,7 @@ import numpy as np
 import pytest
 
 from leb.freeze.datasets import FPDataset
-from leb.freeze.fp import fp_recover, PupilRecoveryMethod, Pupil
+from leb.freeze.fp import fp_recover, FPRecoveryError, PupilRecoveryMethod, Pupil
 
 
 NUM_PX = (64, 64)
@@ -18,6 +18,15 @@ def fake_dataset() -> FPDataset:
 
 
 @pytest.fixture
+def fake_dataset_not_square() -> FPDataset:
+    images = np.zeros((10, NUM_PX[0], NUM_PX[1] - 1), dtype=np.float32)
+    wavevectors = np.zeros((10, 3), dtype=np.float32)
+    led_indexes = np.zeros((10, 2), dtype=np.int32)
+
+    return FPDataset(images, wavevectors, led_indexes)
+
+
+@pytest.fixture
 def fake_pupil() -> Pupil:
     return Pupil.from_system_params(num_px=NUM_PX[0])
 
@@ -26,6 +35,13 @@ def test_fp_recover(fake_dataset, fake_pupil):
     pupil_recovery_method = PupilRecoveryMethod.NONE
 
     fp_recover(fake_dataset, fake_pupil, pupil_recovery_method=pupil_recovery_method)
+
+
+def test_fp_recover_images_must_be_square(fake_dataset_not_square, fake_pupil):
+    pupil_recovery_method = PupilRecoveryMethod.NONE
+
+    with pytest.raises(FPRecoveryError):
+        fp_recover(fake_dataset_not_square, fake_pupil, pupil_recovery_method=pupil_recovery_method)
 
 
 def test_aberrated_pupil_should_not_fail_if_missing_noll_indexes():